cpufreq: exynos: iks: be consistent about boot frequency
The IKS cpufreq driver keeps track of boot frequency in two separate
ways: in the per-cluster exynos_dvfs_info struct and the global
boot_freq variable.  The two are used interchangeably, though the
global boot_freq is updated on resume.  If the firmware resumes the
system at a frequency different than the boot frequency, this will
cause us to suspend at the wrong frequency on subsequent suspends.
Get rid of the global boot_freq variable so we always suspend at the
same frequency and it doesn't matter if we resume at something differnt.

BUG=chrome-os-partner:21430
TEST=suspend_stress_test -c 25; observe that there are no failures
and no imprecise external aborts.

Change-Id: I09c6b468e897ad09cde5795fca672b9d7c5b64af
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64700
Reviewed-by: Doug Anderson <dianders@chromium.org>