CHROMIUM: drm: Refactor drm_bridge to track upstream patch
This patch changes drm_bridge such that it is now statically
associated with an encoder instead of a connector. This requires
the bridge to be connected to the encoder statically on init.

This patch renames prepare to disable and adds new hooks for
post_disable, mode_set, pre_enable, enable which are called from crtc
helper's mode_set function.

Drop this patch in the next version of the kernel in favor of "drm: Add
drm_bridge"

BUG=chromium:266008
TEST=Tested on snow

Change-Id: Ie78cf82253b77d660b04220b3f901a35ca10b41d
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/64778