ARM: exynos: Partial revert of "HACK: ARM: exynos: WiFi power-on sequence"
For pit we ended up with a (marginally) cleaner solution for powering
on the WiFi module by tying the regulator straight into the mmc
device.  Make all devices the same in kernel-3.8 and have them all use
the same solution.

This is a partial revert of (2da402a HACK: ARM: exynos: WiFi power-on
sequence), AKA <https://gerrit.chromium.org/gerrit/47479>.

BUG=chrome-os-partner:20221
TEST=Boot up snow and make sure WiFi works.
TEST=Boot up snow and do a suspend/resume.  WiFi still works.

Change-Id: I97e5538f74318e67bfd3972c4f75455eed0121aa
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/64850
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>