CHROMIUM: s5p-mfc: don't leak private buffers and cause havoc on deallocating them
Releasing buffers via REQBUFS(0) results in the driver allocating over
existing buffers and leaking memory, so check if we already have something
allocated and free it before proceeding with allocation.

Also don't overwrite sizes of already allocated buffers, so that we don't
randomly explode if we fail allocating fresh ones and then try to free
the old ones with sizes of buffers that failed to allocate.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chromium:269460
TEST=boot, video playback

Change-Id: I24919a65e9d5dfb1e096a87fe208ac94180bcc63
Reviewed-on: https://gerrit.chromium.org/gerrit/64946
Tested-by: Pawel Osciak <posciak@chromium.org>
Reviewed-by: kiran <kiran@chromium.org>
Reviewed-by: John Sheu <sheu@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>