CHROMIUM: exynos: hdmi: Dereference ddc_node on use.
Order of initialization of exynos_hdmi and ddc is not currently guaranteed.
This patch allows the DDC channel to be used even if it happens to be
initialized after hdmi.

This patch should be dropped when HDMI driver initialization is guaranteed to
happen after DDC channel on which it depends.

BUG=chrome-os-partner:18395
TEST=Use external HDMI port on spring and observe that it still works.

Change-Id: Iecbe80fa1a37fd83b2c31eac3c0b7e5284b8c033
Signed-off-by: Jeremy Thorpe <jeremyt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65014