CHROMIUM: hid: quickstep: Initial add of driver
The ChomeOS Touch team uses a usb device built in-house to measure the
drag latency on touchpads.  This device communicates with the OS over
USB using a custom driver.  This CL adds that driver so that by default
Chromebooks will have the driver ready.  Previously, the kernel had to
be rebuilt for every device we wanted to test every time, so it will
save a lot of time on our end to already have it in there.

The device (named Quickstep) consists of a laser that lays across the
touchpad and hits a phototransistor on the other side.  The device
reports over USB HIDRAW a 1 or a 0 immediates each time the laser is
broken/unbroken by a finger inbetween the sensor and the laser.  This
driver simply listens for these events and records the time at which
they occur.  These events are stored in memory and can be accessed via
sysfs entries for later analysis by the latency-measuring scripts.

BUG=chromium:328451
TEST=Rebuilt kernel and used with the device.  Was able to access its
sysfs entries to both collect data and clear the logs.

Change-Id: I649c8625275bbdd68a301b96461c7b57c8f7ac8f
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179963
Reviewed-by: Olof Johansson <olofj@chromium.org>