CHROMIUM: drm/bridge: ps8622: allow use of external PWM
The ps8622/ps8625 have an internal PWM which is used by default to control
backlight brightness.  It may be desirable, however, to pass-through an
external PWM which has a higher resolution.  Add the device-tree property
"use-external-pwm" which indicates that the bridge should be configured
to use the external PWM.

BUG=chrome-os-partner:19401
TEST=Once DT bindings are added; observe that the backlight is controlled
using the bl_pwm driver instead of the ps8622 driver.

Change-Id: Iabee5206b198fbce03fe2f0bf72c8dc568c9515e
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65095
Reviewed-by: Olof Johansson <olofj@chromium.org>