pwm-backlight: add "max-brightness" property
Specifying each individual brightness value via the "brightness-levels"
property can be a pain if we want to use a large continuous range of
brightness values.  Add the property "max-brightness", which can be
given in place of "brightness-levels", that specifies that all values
between 0 and the given value can be used.

BUG=chrome-os-partner:19401
TEST=Once DT bindings are added, observe that we can set any value
from 0 to the maximum value given in the device tree as the brightness:
localhost ~ # cat /sys/class/backlight/backlight.9/max_brightness
2800
localhost ~ # echo 1000 > /sys/class/backlight/backlight.9/brightness

Change-Id: I64c41a163ac3b0fa47f644575bde6cdf2189d343
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65097
Reviewed-by: Olof Johansson <olofj@chromium.org>