drm: exynos : restore lost configuration on dpms off
In the case of restart ui or user signng out, all
the configuration is lost in DPMS OFF as a result
for proper functioning of display, the lost
configuration needs to be restored, this patch
restores it.

This patch also  modifies fimd, as in the case
if there are no page flips, the same may be required.

BUG=chrome-os-partner:21758
TEST=suspend_stress_test,hotplug restart ui
     and user sign out with HDMI connected on peach

Change-Id: If09e158016c67212f7505a136c4d288884b141de
Signed-off-by: Shirish S <s.shirish@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/65181
Tested-by: Shirish S <shirish@chromium.org>
Commit-Queue: dongwon.jung <dw3333.jung@samsung.com>
Reviewed-by: dongwon.jung <dw3333.jung@samsung.com>
Tested-by: dongwon.jung <dw3333.jung@samsung.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>