mmw: dw_mmc: Support MMC_POWER_UP / MMC_POWER_OFF
This changes vmmc to be turned on/off by MMC_POWER_UP / MMC_POWER_OFF
of set_ios().  In order to be multislot happy, it also makes vmmc
refcounted per slot (so when all slots turn off we'll turn vmmc off).

The vmmc power starts out off.  This assumes that vmmc power is
different than vqmmc power so that we can still see the card detect
line going low when the card is inserted.  If the two lines are the
same, we should mark "vmmc" power as an always-on regulator until we
actually add a "vqmmc" attribute for dw_mmc.

BUG=chrome-os-partner:21479
TEST=With series, suspend_stress_test and WiFi still work.
TEST=With series, Marvell WiFi reset trick works.
TEST=With series, suspend_stress_test works booted from SD or eMMC.

Change-Id: Iee68364e56e65d0657cc8cc212964e486990158d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65263
Reviewed-by: Abhilash Kesavan <a.kesavan@samsung.com>