clk: exynos5420: Change the PMS values to save power
Change the PMS values of 300MHz, 600MHz and 1200MHz
levels of PLLs to save power.
Although the values are out of Fvco range,
the Error ratio is 0%.

BUG=chrome-os-partner:20827
TEST=Boot kernel on Peach and run WebGL, Video.
See that the performance is not regressed.
Measure power and see marginal save on ARM, KFC and G3D.
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>

Change-Id: I881bbba7ca21c125ce5a30a9ba20ae749a46c330
Reviewed-on: https://gerrit.chromium.org/gerrit/65501
Tested-by: Arjun.K.V <arjun.kv@samsung.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Arjun.K.V <arjun.kv@samsung.com>