ARM: SAMSUNG: Fix the serial driver initialization in the decompression stub.
In the Samsung platform driver for the UART, there's code which is supposed
to force the UART into FIFO mode, but all it really does is reset the FIFOs.
This change adds some code which actually turns on the FIFOs and does some
other basic setup.

BUG=chrome-os-partner:19420
TEST=Built and booted on pit without initializing the UART in firmware. With
this change the kernel successfully boots. Before this change, it would hang
in the decompression stub.

Change-Id: Ibcd0a9f165dcee2ae74254ee43ddaa61becfdd20
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/65305
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>