CHROMIUM: drm/anx7808: Initialize HDMI status earlier.
It appears that setting HDMI_MODE register bit at output configuration time
sometimes causes audio not to work, through a mechanism that is not clear.
Once downstream hotplug is detected, it is safe to st this register, so let's
do it right after bridge input signal is detected.

BUG=chrome-os-partner:21784
TEST=With external monitor (COBY LEDTV3216) connected through HDMI dongle, plug
in the dongle to the HDMI port.  Play a video with sound and hear sound from
the monitor.

Test 2: With board attached to HDMI dongle and Quantum 980 analyzer, do a
capture, and see that infoframes are transmitted, regardless of whether dongle
or downstream HDMI is plugged in first.

Change-Id: I3080a14cd3aa0cb8c02b182208dcf6b678e3276a
Signed-off-by: Jeremy Thorpe <jeremyt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66530
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>