drm/exynos: Don't adjust the clock in mode_fixup
This patch removes the clock adjustment in mode_fixup from fimd. Adding
the clock after the fact causes the mode published via
connector->get_modes() to have a different clock from the currently
configured mode on the crtc. As such, every time userspace sent it back
to us in set_crtc, we'll notice the difference and do an unnecessary full
modeset.

BUG=chromium:266008
TEST=Tested on snow, noticed the number of modesets on boot went from 18 to 1

Change-Id: Iacf0290e9bdf75cc4f8919f986b663692f84d4fd
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168060