drm/prime: shrink critical section protected by prime lock
When exporting a gem object as a dma-buf the critical section for the
per-fd prime lock is just the adding (and in case of errors, removing)
of the handle to the per-fd lookup cache.

So restrict the critical section to just that part of the function.

This simplifies later reordering.

Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Dave Airlie <airlied@redhat.com>
(cherry picked from commit bdf655de47b0d17ee2efc3bea5f617445ff77adc)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

BUG=chromium:340541
TEST=builds and runs on link and daisy

Change-Id: Ifec577daabdd26fa2318a5ad31c88a5511e4ab97
Reviewed-on: https://chromium-review.googlesource.com/190642
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>