UPSTREAM: usb: xhci: Disable runtime PM suspend for quirky controllers
If a USB controller with XHCI_RESET_ON_RESUME goes to runtime suspend,
a reset will be performed upon runtime resume. Any previously suspended
devices attached to the controller will be re-enumerated at this time.
This will cause problems, for example, if an open system call on the
device triggered the resume (the open call will fail).

Note that this change is only relevant when persist_enabled is not set
for USB devices.

BUG=chromium:239214
TEST=Manual. Verify pm_runtime_get_noresume counter gets incremented for
each device attached to the controller, and decremented on removal.
Also, verify count is kept correctly through suspend / resume.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Change-Id: I993afd7da145ebacd90242939448a29075cc0947
Reviewed-on: https://chromium-review.googlesource.com/66020
Reviewed-by: Julius Werner <jwerner@chromium.org>