CHROMIUM: mmc: dw_mmc: Add non-removable property to the host caps
This add non-removable property to the host capability. There are device
like eMMC which are generally permanently mounted on boards, setting
them as a __removable__ device, cause these device to be removed when
entered suspend. Let the board specify that a card is nonremovable e.g.
eMMC.

BUG=chrome-os-partner:22053
TEST=build and booted on pit/Snow
TEST=suspend_stress_test --memory_check works multiple time

Change-Id: Id84cfbcdb7249393f0f834f24da1667ea83ab0f5
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/66351
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>