drm: exynos_fimd: support all kinds of dithering
The FIMD block in Exynos5420 has the capability to apply the
dithering effect on the video output, and we don't need to rely
on MIE block for the same.

MIE block is needed only on Exynos5250, where in there is no
support for dithering in FIMD.

With this patch, we can select the type of dithering needed
based on the panel present on the baord.

Use the DT property "samsung,use-mie-dithering" to enable MIE block.
Or, Use the DT property "samsung,use-fimd-dithering" to enable
the internal dithering logic of FIMD.

If none of the above 2 properties are specified, dithering
feature will be kept disabled.

This patch also updates the DT bindings documentation
for the newly defined properties.

BUG=chrome-os-partner:11072, chrome-os-partner:22042
TEST=Boot on snow and peach and make sure that color
bands are not seen on the screen.

Change-Id: I0502e0cbf7e4a5a09e78ba196b7c5b01d7c46a04
Signed-off-by: Ajay Kumar <ajaykumar.rs@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/66358
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>