UPSTREAM: net: usb: Add HP hs2434 device to ZLP exception table
This patch adds another entry (HP hs2434 Mobile Broadband) to the list
of exceptional devices that require a zero length packet in order to
function properly. This list was added in commit 844e88f0. The hs2434
is manufactured by Sierra Wireless, who also produces the MC7710,
which the ZLP exception list was created for in the first place. So
hopefully it is just this one producer's devices that will need this
workaround.

Tested on a DM1-4310NR HP notebook, which does not function without this
change.

Signed-off-by: Rob Gardner <robmatic@gmail.com>
Acked-by: Bjørn Mork <bjorn@mork.no>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 03803a59e32453ee5737c6096a295f748f03cc49)

BUG=chromium:326559
TEST=Tested the following:
1. `emerge-peppy chromeos-kernel-next`
2. Manually test cellular support on a DUT with a MBIM modem.
3. Run network_3GIdentifiers, network_3GModemPresent and
   network_3GSmokeTest.

Change-Id: I33e784a6083bd699d0cf0bf39a99519721c3db9c
Signed-off-by: Ben Chan <benchan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187013
Reviewed-by: Paul Stewart <pstew@chromium.org>