arm: exynos: pd: add support for clock nodes in pd device node
While powering on/off a local powerdomain in exynos5 chipsets, the input
clocks to each device are modified. This behaviour is based on the
SYSCLK_SYS_PWR_REG registers.
E.g. SYSCLK_MFC_SYS_PWR_REG = 0x0, the parent of input clock to MFC
				   (aclk333) gets modified to oscclk
			    = 0x1, no change in clocks.
The recommended value of SYSCLK_SYS_PWR_REG before power gating any
domain is 0x0 so we must restore the clocks while powering on a domain
everytime.

This patch adds the framework for getting the required mux and parent clocks
through a power domain device node.
Just setting the parent while powering on is not enough since according
to the clock framework, the parent has never changed. So we set the
parent clock as oscclk before power gating and then set back the correct
parent clock after powering on a domain.

BUG=chrome-os-partner:22253
TEST=Boot on peach. Test Video with S2R and see no drop in performance.

Change-Id: Ia42113d7de7fa7d560c8fe3870922c0b1a887509
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/167194