devfreq: exynos5420: remove aclk300_gscl from devfreq
We get slowdown in video performance when gscl's clock is controlled by
devfreq especially when the video is played in very small browser
windows. Maximizing the window and minimizing it again is enough to get
back the video performance which clearly shows that devfreq is not
setting the adequate frequency to this clock.

BUG=chrome-os-partner:22913
TEST=Run multi instance video playback with suspend stress test. No
slowdown in performance should be observed.

Change-Id: Ifb9df2274d7b8d7a86b9b4c3e3eca5326f7f76f7
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/167465
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>