ARM: dts: Add non-removable property for emmc snow and pit
eMMC is non-removable.  Lets set non-removable property for them so
that we don't waste cycles redetecting after a suspend/resume.

NOTE: We purposely don't add the non-removable property for WiFi
despite it not being removable.  That's because the Marvell driver
uses a trick to reset itself that involves mmc_remove_host() /
mmc_add_host().  That doesn't work if the slot is marked
non-removable.  We still have "keep-power-in-suspend" which does save
a bunch of processing across suspend/resume.

BUG=chrome-os-partner:22220, chrome-os-partner:22053
TEST=build and booted on pit, reset wifi card as below:
     echo 1 > /sys/kernel/debug/mwifiex/mlan0/reset
     Can see wifi get rescaned, connected and working after reseting
     wifi card.
TEST=run suspend_stress_test --memory_check still works

[dianders: based atop reverts to simplify later rebases]

Change-Id: I95eb3e21727f6e8abe601d276f0eb6d3aa0c0713
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/67042