s5p-mfc: call wake_up_dev if in suspend mode
If a frame is still decoding when system enters suspend mode, we wait
on the device queue for a interrupt condition. This sometimes leads to a
timeout because the device queue might not be woken up everytime.
Usually, the context queue gets woken up when that context's frame gets
decoded. This patch adds a condition to wake up the device queue along
with the context queue when the system is in suspend mode.

Since the device queue is now woken up, we don't have to check the
context's int_cond flag while waiting. Also, we can skip calling try_run
after waking up the device queue to ensure that we don't have to wait
for more than one frame to be processed.

BUG=chrome-os-partner:22208
TEST=Boot on peach. Run multi-instance video and suspend stress
test. MFC suspend timeout error should never occur.

Change-Id: I60cb34348c28427bfc332733039d39d5576b5292
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/167462
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>