drm/anx7808: Replace timer with oneshot atomic
This patch replaces the debounce timer in the anx cable detect logic
with a oneshot atomic. The cable detect interrupt should be used just to
notify us that there is a slimport dongle attached to the device. This
ensures we will never have 2 concurrent workers running while also
reducing the time-to-video by one second.

BUG=chrome-os-partner:21538
TEST=Tested on spring

Change-Id: Ia8741e74f9cb1d1f45a62e27809640b8cc5268c3
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/168792