ARM: dts: Remove the SD controller and add in the USB 3.0 controllers.
Kirby doesn't have an SD controller and so shouldn't have one defined in the
device tree.

The USB 3.0 ports need two GPIOs enabled to enable their power, one for the
rail itself, and one for the buffer going to the ports. The kernel expects
there to be one GPIO per port and doesn't allow GPIOs to be shared, but what
we really need is to have a pair of gpios which are both shared between two
ports. Julius suggested that we connect one to each controller which satisfies
both constraint, but that means that if the kernel tries to power down either
port it will power down both. We believe it always controls them in tandem so
this should hopefully not be a problem. It's at least a reasonable first
solution.

BUG=None
TEST=Booted with this device tree on kirby. Plugged in USB 2 and USB 3 keys
and saw that each was recognized. The USB 3 key was recognized as a USB 2
device, but that's a problem for another day. Plugged in a USB network dongle
and saw that an IP was acquired successfully.

Change-Id: Ic89121c8f950038129b34d0b54d8c0d1194c8f19
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168507
Reviewed-by: David Hendrix <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>